+++ executor/builder.go
@@ -30,6 +30,9 @@ import (
 	"github.com/pingcap/errors"
 	"github.com/pingcap/failpoint"
 	"github.com/pingcap/kvproto/pkg/diagnosticspb"
+	"github.com/pingcap/tipb/go-tipb"
+	"go.uber.org/zap"
+
 	"github.com/pingcap/tidb/distsql"
 	"github.com/pingcap/tidb/domain"
 	"github.com/pingcap/tidb/executor/aggfuncs"
@@ -49,6 +52,7 @@ import (
 	"github.com/pingcap/tidb/store/helper"
 	"github.com/pingcap/tidb/table"
 	"github.com/pingcap/tidb/table/tables"
+	util2 "github.com/pingcap/tidb/table/tables/util"
 	"github.com/pingcap/tidb/table/temptable"
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util"
@@ -62,8 +66,6 @@ import (
 	"github.com/pingcap/tidb/util/ranger"
 	"github.com/pingcap/tidb/util/rowcodec"
 	"github.com/pingcap/tidb/util/timeutil"
-	"github.com/pingcap/tipb/go-tipb"
-	"go.uber.org/zap"
 )
 
 var (
@@ -436,7 +438,7 @@ func buildIdxColsConcatHandleCols(tblInfo *model.TableInfo, indexInfo *model.Ind
 	handleLen := 1
 	var pkCols []*model.IndexColumn
 	if tblInfo.IsCommonHandle {
-		pkIdx := tables.FindPrimaryIndex(tblInfo)
+		pkIdx := util2.FindPrimaryIndex(tblInfo)
 		pkCols = pkIdx.Columns
 		handleLen = len(pkIdx.Columns)
 	}
@@ -504,7 +506,7 @@ func buildHandleColsForExec(sctx *stmtctx.StatementContext, tblInfo *model.Table
 			ID:      c.ID,
 		}
 	}
-	pkIdx := tables.FindPrimaryIndex(tblInfo)
+	pkIdx := util2.FindPrimaryIndex(tblInfo)
 	for i, c := range pkIdx.Columns {
 		tblCols[c.Offset].Index = len(idxInfo.Columns) + i
 	}
@@ -1867,7 +1869,7 @@ func (b *executorBuilder) buildUnionAll(v *plannercore.PhysicalUnionAll) Executo
 
 func buildHandleColsForSplit(sc *stmtctx.StatementContext, tbInfo *model.TableInfo) plannercore.HandleCols {
 	if tbInfo.IsCommonHandle {
-		primaryIdx := tables.FindPrimaryIndex(tbInfo)
+		primaryIdx := util2.FindPrimaryIndex(tbInfo)
 		tableCols := make([]*expression.Column, len(tbInfo.Columns))
 		for i, col := range tbInfo.Columns {
 			tableCols[i] = &expression.Column{
@@ -3411,7 +3413,7 @@ func buildNoRangeIndexLookUpReader(b *executorBuilder, v *plannercore.PhysicalIn
 			e.handleIdx = append(e.handleIdx, handleCol.Index)
 		}
 		e.handleCols = v.CommonHandleCols
-		e.primaryKeyIndex = tables.FindPrimaryIndex(tbl.Meta())
+		e.primaryKeyIndex = util2.FindPrimaryIndex(tbl.Meta())
 	}
 	return e, nil
 }
@@ -3553,7 +3555,7 @@ func (b *executorBuilder) buildIndexMergeReader(v *plannercore.PhysicalIndexMerg
 			ret.ranges = append(ret.ranges, v.PartialPlans[i][0].(*plannercore.PhysicalTableScan).Ranges)
 			if ret.table.Meta().IsCommonHandle {
 				tblInfo := ret.table.Meta()
-				sctx.IndexNames = append(sctx.IndexNames, tblInfo.Name.O+":"+tables.FindPrimaryIndex(tblInfo).Name.O)
+				sctx.IndexNames = append(sctx.IndexNames, tblInfo.Name.O+":"+util2.FindPrimaryIndex(tblInfo).Name.O)
 			}
 		}
 	}
