+++ table/tables/tables.go
@@ -44,6 +44,7 @@ import (
 	"github.com/pingcap/tidb/statistics"
 	"github.com/pingcap/tidb/table"
 	context2 "github.com/pingcap/tidb/table/tables/context"
+	util2 "github.com/pingcap/tidb/table/tables/util"
 	"github.com/pingcap/tidb/tablecodec"
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util"
@@ -512,37 +513,9 @@ func adjustRowValuesBuf(writeBufs *variable.WriteStmtBufs, rowLen int) {
 	writeBufs.AddRowValues = writeBufs.AddRowValues[:adjustLen]
 }
 
-// FindPrimaryIndex uses to find primary index in tableInfo.
-func FindPrimaryIndex(tblInfo *model.TableInfo) *model.IndexInfo {
-	var pkIdx *model.IndexInfo
-	for _, idx := range tblInfo.Indices {
-		if idx.Primary {
-			pkIdx = idx
-			break
-		}
-	}
-	return pkIdx
-}
-
-// commonAddRecordKey is used as key in `sessionctx.Context.Value(key)`
-type commonAddRecordKey struct{}
-
-// String implement `stringer.String` for CommonAddRecordKey
-func (c commonAddRecordKey) String() string {
-	return "_common_add_record_context_key"
-}
-
-// AddRecordCtxKey is key in `sessionctx.Context` for CommonAddRecordCtx
-var AddRecordCtxKey = commonAddRecordKey{}
-
-// SetAddRecordCtx set a CommonAddRecordCtx to session context
-func SetAddRecordCtx(ctx sessionctx.Context, r *context2.CommonAddRecordCtx) {
-	ctx.SetValue(AddRecordCtxKey, r)
-}
-
 // ClearAddRecordCtx remove `CommonAddRecordCtx` from session context
 func ClearAddRecordCtx(ctx sessionctx.Context) {
-	ctx.ClearValue(AddRecordCtxKey)
+	ctx.ClearValue(context2.AddRecordCtxKey)
 }
 
 // TryGetCommonPkColumnIds get the IDs of primary key column if the table has common handle.
@@ -550,7 +523,7 @@ func TryGetCommonPkColumnIds(tbl *model.TableInfo) []int64 {
 	if !tbl.IsCommonHandle {
 		return nil
 	}
-	pkIdx := FindPrimaryIndex(tbl)
+	pkIdx := util2.FindPrimaryIndex(tbl)
 	pkColIds := make([]int64, 0, len(pkIdx.Columns))
 	for _, idxCol := range pkIdx.Columns {
 		pkColIds = append(pkColIds, tbl.Columns[idxCol.Offset].ID)
@@ -578,7 +551,7 @@ func TryGetCommonPkColumns(tbl table.Table) []*table.Column {
 	if !tbl.Meta().IsCommonHandle {
 		return nil
 	}
-	pkIdx := FindPrimaryIndex(tbl.Meta())
+	pkIdx := util2.FindPrimaryIndex(tbl.Meta())
 	cols := tbl.Cols()
 	pkCols := make([]*table.Column, 0, len(pkIdx.Columns))
 	for _, idxCol := range pkIdx.Columns {
@@ -664,7 +637,7 @@ func (t *TableCommon) AddRecord(sctx sessionctx.Context, r []types.Datum, opts .
 			recordID = kv.IntHandle(r[tblInfo.GetPkColInfo().Offset].GetInt64())
 			hasRecordID = true
 		} else if tblInfo.IsCommonHandle {
-			pkIdx := FindPrimaryIndex(tblInfo)
+			pkIdx := util2.FindPrimaryIndex(tblInfo)
 			pkDts := make([]types.Datum, 0, len(pkIdx.Columns))
 			for _, idxCol := range pkIdx.Columns {
 				pkDts = append(pkDts, r[idxCol.Offset])
@@ -703,7 +676,7 @@ func (t *TableCommon) AddRecord(sctx sessionctx.Context, r []types.Datum, opts .
 
 	var colIDs, binlogColIDs []int64
 	var row, binlogRow []types.Datum
-	if recordCtx, ok := sctx.Value(AddRecordCtxKey).(*context2.CommonAddRecordCtx); ok {
+	if recordCtx, ok := sctx.Value(context2.AddRecordCtxKey).(*context2.CommonAddRecordCtx); ok {
 		colIDs = recordCtx.ColIDs[:0]
 		row = recordCtx.Row[:0]
 	} else {
@@ -915,119 +888,11 @@ func RowWithCols(t table.Table, ctx sessionctx.Context, h kv.Handle, cols []*tab
 	if err != nil {
 		return nil, err
 	}
-	v, _, err := DecodeRawRowData(ctx, t.Meta(), h, cols, value)
+	v, _, err := util2.DecodeRawRowData(ctx, t.Meta(), h, cols, value)
 	if err != nil {
 		return nil, err
 	}
 	return v, nil
-}
-
-func containFullColInHandle(meta *model.TableInfo, col *table.Column) (containFullCol bool, idxInHandle int) {
-	pkIdx := FindPrimaryIndex(meta)
-	for i, idxCol := range pkIdx.Columns {
-		if meta.Columns[idxCol.Offset].ID == col.ID {
-			idxInHandle = i
-			containFullCol = idxCol.Length == types.UnspecifiedLength
-			return
-		}
-	}
-	return
-}
-
-// DecodeRawRowData decodes raw Row data into a datum slice and a (columnID:columnValue) map.
-func DecodeRawRowData(ctx sessionctx.Context, meta *model.TableInfo, h kv.Handle, cols []*table.Column,
-	value []byte) ([]types.Datum, map[int64]types.Datum, error) {
-	v := make([]types.Datum, len(cols))
-	colTps := make(map[int64]*types.FieldType, len(cols))
-	prefixCols := make(map[int64]struct{})
-	for i, col := range cols {
-		if col == nil {
-			continue
-		}
-		if col.IsPKHandleColumn(meta) {
-			if mysql.HasUnsignedFlag(col.Flag) {
-				v[i].SetUint64(uint64(h.IntValue()))
-			} else {
-				v[i].SetInt64(h.IntValue())
-			}
-			continue
-		}
-		if col.IsCommonHandleColumn(meta) && !types.NeedRestoredData(&col.FieldType) {
-			if containFullCol, idxInHandle := containFullColInHandle(meta, col); containFullCol {
-				dtBytes := h.EncodedCol(idxInHandle)
-				_, dt, err := codec.DecodeOne(dtBytes)
-				if err != nil {
-					return nil, nil, err
-				}
-				dt, err = tablecodec.Unflatten(dt, &col.FieldType, ctx.GetSessionVars().Location())
-				if err != nil {
-					return nil, nil, err
-				}
-				v[i] = dt
-				continue
-			}
-			prefixCols[col.ID] = struct{}{}
-		}
-		colTps[col.ID] = &col.FieldType
-	}
-	rowMap, err := tablecodec.DecodeRowToDatumMap(value, colTps, ctx.GetSessionVars().Location())
-	if err != nil {
-		return nil, rowMap, err
-	}
-	defaultVals := make([]types.Datum, len(cols))
-	for i, col := range cols {
-		if col == nil {
-			continue
-		}
-		if col.IsPKHandleColumn(meta) || (col.IsCommonHandleColumn(meta) && !types.NeedRestoredData(&col.FieldType)) {
-			if _, isPrefix := prefixCols[col.ID]; !isPrefix {
-				continue
-			}
-		}
-		ri, ok := rowMap[col.ID]
-		if ok {
-			v[i] = ri
-			continue
-		}
-		if col.IsGenerated() && !col.GeneratedStored {
-			continue
-		}
-		if col.ChangeStateInfo != nil {
-			v[i], _, err = GetChangingColVal(ctx, cols, col, rowMap, defaultVals)
-		} else {
-			v[i], err = GetColDefaultValue(ctx, col, defaultVals)
-		}
-		if err != nil {
-			return nil, rowMap, err
-		}
-	}
-	return v, rowMap, nil
-}
-
-// GetChangingColVal gets the changing column value when executing "modify/change column" statement.
-// For statement like update-where, it will fetch the old Row out and insert it into kv again.
-// Since update statement can see the writable columns, it is responsible for the casting relative column / get the fault value here.
-// old Row : a-b-[nil]
-// new Row : a-b-[a'/default]
-// Thus the writable new Row is corresponding to Write-Only constraints.
-func GetChangingColVal(ctx sessionctx.Context, cols []*table.Column, col *table.Column, rowMap map[int64]types.Datum, defaultVals []types.Datum) (_ types.Datum, isDefaultVal bool, err error) {
-	relativeCol := cols[col.ChangeStateInfo.DependencyColumnOffset]
-	idxColumnVal, ok := rowMap[relativeCol.ID]
-	if ok {
-		idxColumnVal, err = table.CastValue(ctx, idxColumnVal, col.ColumnInfo, false, false)
-		// TODO: Consider sql_mode and the error msg(encounter this error check whether to rollback).
-		if err != nil {
-			return idxColumnVal, false, errors.Trace(err)
-		}
-		return idxColumnVal, false, nil
-	}
-
-	idxColumnVal, err = GetColDefaultValue(ctx, col, defaultVals)
-	if err != nil {
-		return idxColumnVal, false, errors.Trace(err)
-	}
-
-	return idxColumnVal, true, nil
 }
 
 // RemoveRecord implements table.Table RemoveRecord interface.
@@ -1303,7 +1168,7 @@ func IterRecords(t table.Table, ctx sessionctx.Context, cols []*table.Column,
 				data[col.Offset] = rowMap[col.ID]
 				continue
 			}
-			data[col.Offset], err = GetColDefaultValue(ctx, col, defaultVals)
+			data[col.Offset], err = util2.GetColDefaultValue(ctx, col, defaultVals)
 			if err != nil {
 				return err
 			}
@@ -1338,29 +1203,6 @@ func tryDecodeColumnFromCommonHandle(col *table.Column, handle kv.Handle, pkIds 
 		return d, nil
 	}
 	return types.Datum{}, nil
-}
-
-// GetColDefaultValue gets a column default value.
-// The defaultVals is used to avoid calculating the default value multiple times.
-func GetColDefaultValue(ctx sessionctx.Context, col *table.Column, defaultVals []types.Datum) (
-	colVal types.Datum, err error) {
-	if col.GetOriginDefaultValue() == nil && mysql.HasNotNullFlag(col.Flag) {
-		return colVal, errors.New("Miss column")
-	}
-	if col.State != model.StatePublic {
-		return colVal, nil
-	}
-	if defaultVals[col.Offset].IsNull() {
-		colVal, err = table.GetColOriginDefaultValue(ctx, col.ToInfo())
-		if err != nil {
-			return colVal, err
-		}
-		defaultVals[col.Offset] = colVal
-	} else {
-		colVal = defaultVals[col.Offset]
-	}
-
-	return colVal, nil
 }
 
 // AllocHandle allocate a new handle.
@@ -1483,7 +1325,7 @@ func CanSkip(info *model.TableInfo, col *table.Column, value *types.Datum) bool 
 		return true
 	}
 	if col.IsCommonHandleColumn(info) {
-		pkIdx := FindPrimaryIndex(info)
+		pkIdx := util2.FindPrimaryIndex(info)
 		for _, idxCol := range pkIdx.Columns {
 			if info.Columns[idxCol.Offset].ID != col.ID {
 				continue
@@ -1749,7 +1591,7 @@ func TryGetHandleRestoredDataWrapper(t table.Table, row []types.Datum, rowMap ma
 		return nil
 	}
 	rsData := make([]types.Datum, 0, 4)
-	pkIdx := FindPrimaryIndex(t.Meta())
+	pkIdx := util2.FindPrimaryIndex(t.Meta())
 	for _, pkIdxCol := range pkIdx.Columns {
 		pkCol := t.Meta().Columns[pkIdxCol.Offset]
 		if !types.NeedRestoredData(&pkCol.FieldType) {
